Separated function verifier from alive-tv.cpp
