+++ CMakeLists.txt
@@ -187,6 +187,7 @@ if (BUILD_LLVM_UTILS OR BUILD_TV)
     "llvm_util/llvm_optimizer.cpp"
     "llvm_util/llvm2alive.cpp"
     "llvm_util/utils.cpp"
+    "llvm_util/function_verify.cpp"
   )
 
   add_library(llvm_util STATIC ${LLVM_UTIL_SRCS})
