+++ tools/alive-tv.cpp
@@ -1,39 +1,34 @@
 // Copyright (c) 2018-present The Alive2 Authors.
 // Distributed under the MIT license that can be found in the LICENSE file.
 
-#include "cache/cache.h"
+#include "llvm_util/function_verify.h"
 #include "llvm_util/llvm2alive.h"
 #include "llvm_util/llvm_optimizer.h"
 #include "smt/smt.h"
 #include "tools/transform.h"
 #include "util/version.h"
 
-#include "llvm/ADT/StringExtras.h"
 #include "llvm/ADT/Triple.h"
 #include "llvm/Analysis/TargetLibraryInfo.h"
 #include "llvm/Bitcode/BitcodeReader.h"
-#include "llvm/InitializePasses.h"
 #include "llvm/IR/LLVMContext.h"
-#include "llvm/IR/Function.h"
 #include "llvm/IR/Module.h"
-#include "llvm/IR/LegacyPassManager.h"
 #include "llvm/IRReader/IRReader.h"
 #include "llvm/Passes/PassBuilder.h"
 #include "llvm/Support/PrettyStackTrace.h"
 #include "llvm/Support/Signals.h"
 #include "llvm/Support/SourceMgr.h"
-#include "llvm/Transforms/IPO/PassManagerBuilder.h"
 #include "llvm/Transforms/Utils/Cloning.h"
 
 #include <fstream>
 #include <iostream>
-#include <sstream>
 #include <utility>
 
 using namespace tools;
 using namespace util;
 using namespace std;
 using namespace llvm_util;
+using namespace verifier;
 
 #define LLVM_ARGS_PREFIX ""
 #define ARGS_SRC_TGT
@@ -87,168 +82,7 @@ std::unique_ptr<llvm::Module> openInputFile(llvm::LLVMContext &Context,
   return M;
 }
 
-optional<smt::smt_initializer> smt_init;
-unique_ptr<Cache> cache;
-
-struct Results {
-  Transform t;
-  string error;
-  Errors errs;
-  enum {
-    ERROR,
-    TYPE_CHECKER_FAILED,
-    SYNTACTIC_EQ,
-    CORRECT,
-    UNSOUND,
-    FAILED_TO_PROVE
-  } status;
-
-  static Results Error(string &&err) {
-    Results r;
-    r.status = ERROR;
-    r.error = std::move(err);
-    return r;
-  }
-};
-
-Results verify(llvm::Function &F1, llvm::Function &F2,
-               llvm::TargetLibraryInfoWrapperPass &TLI,
-               bool print_transform = false,
-               bool always_verify = false) {
-  auto fn1 = llvm2alive(F1, TLI.getTLI(F1), true);
-  if (!fn1)
-    return Results::Error("Could not translate '" + F1.getName().str() +
-                          "' to Alive IR\n");
-
-  auto fn2 = llvm2alive(F2, TLI.getTLI(F2), false, fn1->getGlobalVarNames());
-  if (!fn2)
-    return Results::Error("Could not translate '" + F2.getName().str() +
-                          "' to Alive IR\n");
-
-  Results r;
-  r.t.src = std::move(*fn1);
-  r.t.tgt = std::move(*fn2);
-
-  if (!always_verify) {
-    stringstream ss1, ss2;
-    r.t.src.print(ss1);
-    r.t.tgt.print(ss2);
-    if (std::move(ss1).str() == std::move(ss2).str()) {
-      if (print_transform)
-        r.t.print(*out, {});
-      r.status = Results::SYNTACTIC_EQ;
-      return r;
-    }
-  }
-
-  smt_init->reset();
-  r.t.preprocess();
-  TransformVerify verifier(r.t, false);
-
-  if (print_transform)
-    r.t.print(*out, {});
-
-  {
-    auto types = verifier.getTypings();
-    if (!types) {
-      r.status = Results::TYPE_CHECKER_FAILED;
-      return r;
-    }
-    assert(types.hasSingleTyping());
-  }
-
-  r.errs = verifier.verify();
-  if (r.errs) {
-    r.status = r.errs.isUnsound() ? Results::UNSOUND : Results::FAILED_TO_PROVE;
-  } else {
-    r.status = Results::CORRECT;
-  }
-  return r;
-}
-
-unsigned num_correct = 0;
-unsigned num_unsound = 0;
-unsigned num_failed = 0;
-unsigned num_errors = 0;
-
-bool compareFunctions(llvm::Function &F1, llvm::Function &F2,
-                      llvm::TargetLibraryInfoWrapperPass &TLI) {
-  auto r = verify(F1, F2, TLI, !opt_quiet, opt_always_verify);
-  if (r.status == Results::ERROR) {
-    *out << "ERROR: " << r.error;
-    ++num_errors;
-    return true;
-  }
-
-  if (opt_print_dot) {
-    r.t.src.writeDot("src");
-    r.t.tgt.writeDot("tgt");
-  }
-
-  switch (r.status) {
-  case Results::ERROR:
-    UNREACHABLE();
-    break;
-
-  case Results::SYNTACTIC_EQ:
-    *out << "Transformation seems to be correct! (syntactically equal)\n\n";
-    ++num_correct;
-    break;
-
-  case Results::CORRECT:
-    *out << "Transformation seems to be correct!\n\n";
-    ++num_correct;
-    break;
-
-  case Results::TYPE_CHECKER_FAILED:
-    *out << "Transformation doesn't verify!\n"
-            "ERROR: program doesn't type check!\n\n";
-    ++num_errors;
-    return true;
-
-  case Results::UNSOUND:
-    *out << "Transformation doesn't verify!\n\n";
-    if (!opt_quiet)
-      *out << r.errs << endl;
-    ++num_unsound;
-    return false;
-
-  case Results::FAILED_TO_PROVE:
-    *out << r.errs << endl;
-    ++num_failed;
-    return true;
-  }
-
-  if (opt_bidirectional) {
-    r = verify(F2, F1, TLI, false, opt_always_verify);
-    switch (r.status) {
-    case Results::ERROR:
-    case Results::TYPE_CHECKER_FAILED:
-      UNREACHABLE();
-      break;
-
-    case Results::SYNTACTIC_EQ:
-    case Results::CORRECT:
-      *out << "These functions seem to be equivalent!\n\n";
-      return true;
-
-    case Results::FAILED_TO_PROVE:
-      *out << "Failed to verify the reverse transformation\n\n";
-      if (!opt_quiet)
-        *out << r.errs << endl;
-      return true;
-
-    case Results::UNSOUND:
-      *out << "Reverse transformation doesn't verify!\n\n";
-      if (!opt_quiet)
-        *out << r.errs << endl;
-      return false;
-    }
-  }
-  return true;
-}
-
-llvm::Function *findFunction(llvm::Module &M, const string &FName) {
+llvm::Function *findFunction(llvm::Module &M, const std::string &FName) {
   for (auto &F : M) {
     if (F.isDeclaration())
       continue;
@@ -259,7 +93,6 @@ llvm::Function *findFunction(llvm::Module &M, const string &FName) {
   return 0;
 }
 }
-
 
 int main(int argc, char **argv) {
   llvm::sys::PrintStackTraceOnErrorSignal(argv[0]);
@@ -313,6 +146,9 @@ convenient way to demonstrate an existing optimizer bug.
 
   llvm_util::initializer llvm_util_init(*out, DL);
   smt_init.emplace();
+
+  initialize_verifier(opt_quiet, opt_always_verify, opt_print_dot,
+                      opt_bidirectional, out);
 
   unique_ptr<llvm::Module> M2;
   if (opt_file2.empty()) {
